minor tweaks, updated MANGLE interface

+++ lib/minimal_mangle.c
@@ -193,7 +193,7 @@ mply_pix_index_from_id( MANGLE_PLY const *const ply, MANGLE_INT id )
     size_t npix = mply_pix_count( ply->pix_res );
     index = id - mply_pix_id_start( ply );
 
-    if( index < 0 || index >= npix ) {
+    if( index < 0 || (size_t) index >= npix ) {
         fprintf( stderr,
                  "MANGLE Error: pixel_id=%zd does not match with pixel_res=%zd.\n",
                  ( ssize_t ) id, ( ssize_t ) ply->pix_res );
